+++ style-processor.js
@@ -1,241 +1,115 @@
-'use strict';
+'use strict'
 
-var CachingWriter       = require('broccoli-caching-writer'),
-    merge               = require('lodash.merge'),
-    deepMerge           = require('deepmerge'),
-    mkdirp              = require('mkdirp'),
-    path                = require('path'),
-    fs                  = require('fs'),
-    includePathSearcher = require('include-path-searcher'),
+const CachingWriter = require('broccoli-caching-writer')
 
-    //Preprocessors
-    nodeSass            = require('node-sass'),
-    postcss             = require('postcss'),
-    less                = require('less');
+const mkdirp = require('mkdirp-promise')
+const path = require('path')
+const fs = require('fs-promise')
+const co = require('co')
 
-StyleProcessor.prototype = Object.create(CachingWriter.prototype);
-StyleProcessor.prototype.constructor = StyleProcessor;
+const isGlob = require('is-glob')
+const minimatch = require('minimatch')
 
+const loadProcessor = require('./load-processor')
+
+StyleProcessor.prototype = Object.create(CachingWriter.prototype)
+StyleProcessor.prototype.constructor = StyleProcessor
+
+/**
+ * @class StyleProcessor
+ * @extends CachingWriter
+ */
 function StyleProcessor(inputNodes, inputFile, outputFile, _options) {
 
-    if (!(this instanceof StyleProcessor)) {
-        return new StyleProcessor(sourceTrees, inputFile, outputFile, _options);
-    }
+	if(!(this instanceof StyleProcessor)) {
+		return new StyleProcessor(inputNodes, inputFile, outputFile, _options)
+	}
 
-    CachingWriter.call(this, inputNodes, {
-        annotation: _options.annotation
-    });
+	CachingWriter.call(this, inputNodes, {
+		annotation: _options.annotation
+	})
 
-    this.options        = merge({}, _options);
-    this.inputFile      = inputFile;
-    this.outputFile     = outputFile;
+	this._projectRoot = _options.projectRoot
+	this._processors = _options.processors
 
-    //Import path for preprocessors that allow including other files
-    this.importPath = '.' + path.dirname(inputFile);
+	this._inputFilePath = `.${inputFile}`
+	this._inputFileDir = path.dirname(this._inputFilePath)
+	this._inputFileName = path.relative(`${_options.projectRoot}/app/styles`, this._inputFilePath) // relative path
+	this._outputFile = outputFile
+
+	//Information which is passed to every single style processor
+	this._fileInfo = {
+		inputFile: this._inputFilePath,
+		importPath: this._inputFileDir,
+		extension: _options.ext
+	}
 }
 
-StyleProcessor.prototype.build = function() {
-    var self = this;
 
-    //Create destination file
-    var destFile    = path.join(this.outputPath, this.outputFile);
+/**
+ * Handles the fileContent transformation
+ *
+ * @returns {Promise}
+ */
+StyleProcessor.prototype.build = co.wrap(function*() {
+	let fileContents = yield fs.readFile(this._inputFilePath, {
+		encoding: 'utf8'
+	})
 
-    //Read input file
-    var filename    = includePathSearcher.findFileSync(this.inputFile, this.inputPaths);
-    var dataRaw     = fs.readFileSync(filename, 'utf8');
+	for (let processor of this._processors) {
 
-    //Execute processors in sequential order
-    return this.options.processors.reduce(function(currentPromise, nextProcessor) {
-
-        //Chain em all
-        return currentPromise
-
-        .then(function(dataProcessed) {
-            return self.getProcessorPromise(dataProcessed, nextProcessor, filename);
-        })
-
-        //Catch Error from preprocessor and pass it one level higher!
-        .catch(function(errBroccoli) {
-            return Promise.reject(errBroccoli);
-        });
-
-    }, Promise.resolve(dataRaw))
-    // Stylesheet preprocessing complete -> Write to file
-    .then(function(dataProcessed) {
-        mkdirp.sync(path.dirname(destFile));
-
-        fs.writeFileSync(destFile, dataProcessed, {
-            encoding: 'utf8'
-        });
-
-        //Resolve to tell broccoli we finished our async stuff
-        return Promise.resolve();
-    })
-    // Stylesheet preprocessing failed -> Throw an error
-    .catch(function(errBroccoli) {
-        //Add missing data to error
-        errBroccoli.file = this.inputFile;
-        errBroccoli.treeDir = this.inputPaths[0];
-
-        //Reject and tell broccoli we got some error
-        return Promise.reject(errBroccoli);
-    }.bind(this));
-}
-
-StyleProcessor.prototype.getProcessorPromise = function(dataToProcess, processor, currFilename) {
-
-	//chech if the current processor is should process the current data
-	var boolProcess = false;
-	//TODO: why are the filenames fucked up?
-	var needle = '/styles/';
-	var needleIndex = currFilename.indexOf(needle);
-	currFilename = currFilename.substr(needleIndex + needle.length);
-
-	if(typeof processor.file === 'undefined') {
-		boolProcess = true;
-	} else if (processor.file === currFilename) {
-		boolProcess = true;
-	} else if (typeof processor.file === 'array') {
-		if(processor.file.indexOf(currFilename) != -1) {
-			boolProcess = true;
+		if(this._checkProcess(processor.filter)) {
+			const _processor = loadProcessor(processor.type)
+			fileContents = yield _processor(fileContents, processor, this._fileInfo)
 		}
 	}
 
-	if(!boolProcess) {
-		return new Promise(function(res, rej) {
-			return res(dataToProcess);
-		});
+	const outputFile = path.join(this.outputPath, this._outputFile)
+	const outputFileDir = path.dirname(outputFile)
+
+	yield mkdirp(outputFileDir)
+	yield fs.writeFile(outputFile, fileContents, {
+		encoding: 'utf8'
+	})
+
+	return Promise.resolve()
+})
+
+/**
+ * Checks if the file should be processed by a processor filter
+ *
+ * @param {array|string} filter Array of String or String. Supports glob pattern.
+ * @returns {boolean} process Process or not process file
+ * @private
+ */
+StyleProcessor.prototype._checkProcess = function(filter) {
+
+	if (!filter) {
+		return true
 	}
 
-    if(processor.type === 'sass') {
-        return this.compileSass(dataToProcess, processor);
-    } else if (processor.type === 'postcss') {
-        return this.compilePostCSS(dataToProcess, processor, currFilename);
-    } else if(processor.type === 'less') {
-        return this.compileLess(dataToProcess, processor);
-    } else {
-        throw new Error('Unknown processor type');
-    }
+	const fileName = this._inputFileName
+
+	if (isGlob(filter)) {
+		return minimatch(fileName, filter)
+	}
+
+	if (typeof filter === 'string') {
+		return (fileName === filter)
+	}
+
+	if (filter instanceof Array) {
+
+		let _process = false
+
+		for (let _filter of filter) {
+			_process = this._checkProcess(_filter)
+		}
+
+		return _process
+	}
+
+	return false
 }
 
-StyleProcessor.prototype.compileSass = function(data, processor) {
-    return new Promise(function(res, rej) {
-
-        var optionsDefault = {
-            data: data,
-            includePaths: [this.importPath]
-        };
-
-        var optionsConfig = processor.options || {};
-
-        var options = deepMerge(optionsConfig, optionsDefault);
-
-        nodeSass.render(options, function(errSass, result) {
-
-            if(errSass) {
-                //Transform sass error to broccoli error
-                var errBroccoli     = new Error(errSass.message);
-                errBroccoli.line    = errSass.line;
-                errBroccoli.column  = errSass.column;
-
-                return rej(errBroccoli)
-            }
-
-            return res(result.css);
-        });
-    }.bind(this))
-}
-
-StyleProcessor.prototype.compilePostCSS = function(data, processor, filename) {
-
-    if(!processor.plugins) {
-        throw new Error('Please add plugins to your postcss-process!');
-    }
-
-	var processOptions = {
-		//We need this for some plugins (e.q. precss to find imports)
-		from: 'app/styles/' + filename,
-		to: 'styles/' //TODO: What do with this?
-	};
-
-    if(processor.parser) {
-		processOptions.parser = processor.parser;
-    }
-
-	if(processor.syntax) {
-		processOptions.syntax = processor.syntax;
-    }
-
-    //Wrap options in module
-    var postcssPlugins = processor.plugins.reduce(function(pluginArray, curPluginConf) {
-
-        if(typeof curPluginConf.module != 'function') {
-            throw new Error('One of your postcss plugins is not a module!');
-        }
-
-        var pluginOptions = curPluginConf.options || {};
-        pluginArray.push(curPluginConf.module(pluginOptions));
-
-        return pluginArray;
-    }, []);
-
-    return new Promise(function(res, rej) {
-        postcss(postcssPlugins)
-        .process(data, processOptions)
-
-        .then(function(dataProcessed) {
-            return res(dataProcessed.css);
-        })
-
-        .catch(function(errPostCss) {
-            //Transform postcss error to broccoli error
-			//TODO: Do postcss plugins all have different errorMessage properties?
-			var errBroccoli= new Error(errPostCss.message ? errPostCss.message : errPostCss.originalMessage);
-			errBroccoli.line    = errPostCss.lineNumber;
-			errBroccoli.column  = errPostCss.columnNumber;
-
-            return rej(errBroccoli);
-        });
-    });
-}
-
-StyleProcessor.prototype.compileLess = function(data, processor) {
-
-    var optionsConfig   = processor.options || {},
-        optionsDefault  = { plugins: [], paths: [this.importPath] },
-        options         = null;
-
-    //Load plugins
-    if(processor.plugins) {
-        optionsDefault.plugins = processor.plugins.reduce(function(plugins, plugin) {
-            var pluginOptions = plugin.options || null;
-            plugins.push(new plugin.module(pluginOptions));
-            return plugins;
-        }, []);
-    }
-
-    options = merge({}, optionsDefault, optionsConfig);
-
-    options = deepMerge(options, optionsDefault);
-
-    return new Promise(function(res, rej) {
-        less.render(data, options, function(errLess, result) {
-            if(errLess) {
-
-                //Concat error message
-                var errMessage      = errLess.message + "\n" + errLess.extract[0];
-
-                //Transform postcss error to broccoli error
-                var errBroccoli     = new Error(errMessage);
-                errBroccoli.line    = errLess.line;
-                errBroccoli.column  = errLess.column;
-
-                return rej(errBroccoli);
-            }
-
-            return res(result.css);
-        })
-    }.bind(this));
-}
-
-module.exports = StyleProcessor;
+module.exports = StyleProcessor