+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-## 0.0.8
+# 1.0.0
+* Refactors StyleProcessor
+* Processor modules are no longer dependencies and must be installed manually
+* Removes many semicolons ;)
+* Adds eslint
+* Changes file to filter
+* Implements glob pattern for filter
+* Fixes node-sass sourcemap file
+
+
+# 0.0.8
 * Fixes filepath which is passed to postcss